プロジェクト構造の整合性チェック
