+++ .idea/compiler.xml
@@ -9,11 +9,9 @@
         <outputRelativeToContentRoot value="true" />
         <module name="TrainingList (1)" />
         <module name="TrainingList (2)" />
+        <module name="TrainingList" />
       </profile>
     </annotationProcessing>
-    <bytecodeTargetLevel>
-      <module name="TrainingList" target="17" />
-    </bytecodeTargetLevel>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
