--- TrainingList/src/main/java/jp/crossabilitys/work/TrainingList/repository/TrainingRepository.java
@@ -1,16 +1,0 @@
-package jp.crossabilitys.work.TrainingList.repository;
-
-import jp.crossabilitys.work.TrainingList.Entity.TrainingInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-
-/**
- * 訓練情報 Repository
- */
-public interface TrainingRepository extends JpaRepository<TrainingInfo,Long>{
-    @Query(value="SELECT * FROM TRAININGINFO where deleteflg=:deleteflg",nativeQuery = true)
-    List<TrainingInfo> searchAll(boolean deleteflg);
-
-}
