--- TrainingList/src/main/resources/mock/TeacherInfo.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-<head>
-	<!-- Required meta tags -->
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<!-- Bootstrap CSS -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-	<link rel="stylesheet" href="init.css">
-
-	<title>講師割当アプリケーション[講師情報]</title>
-</head>
-
-<body>
-	<div class="border col-5">
-		<br>
-		<h2>講師情報</h2>
-		<br>
-		<div class="row">
-			<div class="col-md">
-				<form>
-					<!--div class="input-group mb-3">
-						<span class="input-group-text" id="basic-addon1">ID</span>
-						<input type="text" class="form-control" value="00001" disabled>
-					</div-->
-					<div class="form-group">
-						<label>ID：</label>
-						<input type="text" class="form-control" value="00001" disabled>
-					</div>
-					<div class="form-group">
-						<label>講師名：</label>
-						<input type="text" class="form-control" placeholder="テスト　太郎">
-					</div>
-					<div class="form-group">
-						<label>担当講座：</label>
-						<select name="TrainingSelect"  class="form-select" aria-label="Default select example">
-							<option value="Training1">〇〇〇初級科</option>
-							<option value="Training2">〇〇〇中級科</option>
-							<option value="Training3">〇〇〇上級科</option>
-						</select>
-					</div>
-				</form>
-			</div>
-		</div>
-		<br>
-		<div class="row center-block text-center">
-			<div class="col-18">
-				<button type="button" class="btn btn-primary btn-block" onclick="location.href='./TeacherList.html'">登録</button>
-				<button type="button" class="btn btn-outline-primary btn-block" onclick="location.href='./TeacherList.html'">キャンセル</button>
-			</div>
-		</div>
-        <div class="col-1">
-        </div>
-    </div>
-
-
-<!-- Optional JavaScript; choose one of the two! -->
-
-    <!-- Option 1: Bootstrap Bundle with Popper -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    <!-- Option 2: Separate Popper and Bootstrap JS -->
-    <!--
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-    -->
-
-
-</body>
-
-</html>
