--- TrainingList/src/main/resources/mock/TeacherSchedule.html
@@ -1,290 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-<head>
-	<!-- Required meta tags -->
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<!-- Bootstrap CSS -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-	<link rel="stylesheet" href="init.css">
-
-	<title>講師割当アプリケーション[講師スケジュール]</title>
-</head>
-
-<body>
-	<div class="col-6">
-		<br>
-		<h4>講師スケジュール</h4>
-		<br>
-		<h5 align="right">講師名：テスト　太郎</h5>
-		<div class="row center-block text-center">
-			<h2><a href= "TeacherSchedule.html" title="前月へ遷移予定。"><</a> 9月 <a href= "TeacherSchedule.html"  title="次月へ遷移予定。">></a></h2>
-		</div>
-		<div  class="row center-block text-center" id="calendar">
-			<!--div class="col-md"-->
-		<table>
-			<tr>
-				<th>日</th>
-				<th>月</th>
-				<th>火</th>
-				<th>水</th>
-				<th>木</th>
-				<th>金</th>
-				<th>土</th>
-			</tr>
-			<tr class="day">
-				<td></td>
-				<td></td>
-				<td></td>
-				<td></td>
-				<td></td>
-				<td>1<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="holiday">2<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-			</tr>
-			<tr class="day">
-				<td class="holiday">3<br>
-					<select  disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>4<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>5<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>6<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>7<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>8<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="holiday">9<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-			</tr>
-			<tr class="day">
-				<td class="holiday">10<br>
-					<select  disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>11<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>12<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>13<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>14<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>15<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="holiday">16<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-			</tr>
-			<tr class="day">
-				<td class="holiday">17<br>
-					<select  disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="nholiday">18<br>
-					<select disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>19<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>20<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>21<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>22<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="nholiday">23<br>
-					<select disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				</tr>
-			<tr class="day">
-				<td class="holiday">24<br>
-					<select  disabled>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>25<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>26<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>27<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>28<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td>29<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2" selected>〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-				<td class="holiday">30<br>
-					<select>
-					<option value="1">―</option>
-					<option value="2">〇</option>
-					<option value="3">×</option>
-					</select>
-				</td>
-			</tr>
-		</table>
-			<!--/div-->
-		</div>
-		<br>
-		<div class="row center-block text-center">
-			<div class="col-18">
-				<button type="button" class="btn btn-primary btn-block" onclick="location.href='./TeacherList.html'">登録</button>
-				<button type="button" class="btn btn-outline-primary btn-block" onclick="location.href='./TeacherList.html'">キャンセル</button>
-			</div>
-		</div>
-		<br>
-        <div class="col-1">
-        </div>
-    </div>
-
-
-<!-- Optional JavaScript; choose one of the two! -->
-
-    <!-- Option 1: Bootstrap Bundle with Popper -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    <!-- Option 2: Separate Popper and Bootstrap JS -->
-    <!--
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-    -->
-
-
-</body>
-
-</html>
