--- TrainingList/src/main/resources/mock/TrainingInfo.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-<head>
-	<!-- Required meta tags -->
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<!-- Bootstrap CSS -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-	<link rel="stylesheet" href="init.css">
-
-	<title>講師割当アプリケーション[訓練情報]</title>
-</head>
-
-<body>
-	<div class="border col-5">
-		<br>
-		<h2>訓練情報</h2>
-		<br>
-		<div class="row">
-			<div class="col-md">
-				<form>
-					<div class="input-group mb-3">
-						<span class="input-group-text">ID</span>
-						<input type="text" class="form-control" value="00001" disabled>
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">訓練名</span>
-						<input type="text" class="form-control" placeholder="〇〇〇初級科" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">募集期間</span>
-						<input type="date" class="form-control" />　～　
-						<input type="date" class="form-control" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">選考日</span>
-						<input type="date" class="form-control" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">合格発表</span>
-						<input type="date" class="form-control" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">訓練期間</span>
-						<input type="date" class="form-control" />　～　
-						<input type="date" class="form-control" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">授業時間</span>
-						<input type="time" class="form-control" />　～　
-						<input type="time" class="form-control" />
-					</div>
-					<div class="input-group mb-3">
-						<span class="input-group-text">授業時間数</span>
-						<input type="number" class="form-control" min="0" max="8" />
-					</div>
-				</form>
-			</div>
-		</div>
-		<br>
-		<div class="row center-block text-center">
-			<div class="col-18">
-				<button type="button" class="btn btn-primary btn-block" onclick="location.href='./TrainingList.html'">登録</button>
-				<button type="button" class="btn btn-outline-primary btn-block" onclick="location.href='./TrainingList.html'">キャンセル</button>
-			</div>
-		</div>
-        <div class="col-1">
-        </div>
-    </div>
-
-
-<!-- Optional JavaScript; choose one of the two! -->
-
-    <!-- Option 1: Bootstrap Bundle with Popper -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-    <!-- Option 2: Separate Popper and Bootstrap JS -->
-    <!--
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-    -->
-
-
-</body>
-
-</html>
