--- TrainingList/src/main/resources/mock/init.css
@@ -1,40 +1,0 @@
-/* ----- 訓練時間表 ----- */
-/* 曜日行 */
-tr.row_dow{
-/* 	height: 20px; */
-}
-/* データ行 */
-tr.row_data{
- 	height: 35px;
-	font-size: x-small; 
-	text-align:center; 
-}
-/* 曜日セル */
-td.dow{
- 	height: 10px;
-	font-size: small;
-	text-align:center; 
-}
-/* 休日 */
-td.holiday{
-	background-color: #dddddd; 
-}
-/* 祝日 */
-td.nholiday{
-	background-color: #ffcccc; 
-}
-/* 講師登録なし */
-td.nodata{
-	background-color: #ff0000; 
-}
-
-/* ----- スケジュール情報 ----- */
-#calendar td.holiday{
-	background-color: #dddddd;
-}
-#calendar td.nholiday{
-	background-color: #ffcccc;
-}
-.inputhours{
-    width:40px;
-}
