--- TrainingList/src/main/resources/static/init.css
@@ -1,103 +1,0 @@
-/* ----- 訓練時間表 ----- */
-/* 曜日行 */
-tr.row_dow{
-/* 	height: 20px; */
-}
-/* データ行 */
-tr.row_data{
- 	height: 35px;
-	font-size: x-small;
-	text-align:center;
-}
-
-/* 曜日セル */
-td.dow{
- 	height: 10px;
-	font-size: small;
-	text-align:center;
-}
-/* 休日 */
-td.holiday{
-	background-color: #d1d1d1;
-}
-/* 祝日 */
-td.nholiday{
-	background-color: #ffcccc;
-}
-/* 講師登録なし */
-td.nodata{
-	background-color: #ff0000;
-}
-td.sHoliday{
-    background-color: #00ffd1;
-}
-/* 対象外 */
-td.outofdata {
-    background-color: #e2e7eb;
-}
-/* コンテキストメニュー */
-#contextmenu{
-    display:none;
-    position:fixed;
-    left:0px;
-    top:0px;
-    padding-top: 0.75em;
-    padding-right: 0.5em;
-    border:1px solid #000;
-    border-radius: 20px;
-    color: #333333;
-    background-color: rgba(255, 255, 255, 0.7);
-}
-#contextmenu ul{
-    list-style:none;
-    padding-left:0.5em;
-}
-
-li:hover {
-    cursor:pointer;
-    color: #ffffff;
-    background-color: dimgray;
-}
-
-/* ----- スケジュール情報 ----- */
-#calendar td.holiday{
-	background-color: #dddddd;
-}
-#calendar td.nholiday{
-	background-color: #ffcccc;
-}
-.inputhours{
-    width:40px;
-}
-/* ツールチップ背景と枠 */
-.tooltip_background{
-    z-index: 1;
-    padding: 0.5em 0.5em;
-    border: solid 1px #000000;
-    position: fixed;
-}
-
-/* ツールチップのデータ部分 */
-.tooltip{
-    z-index: 3;
-    position: relative;
-    opacity: 1;
-}
-
-/*　ツールチップ背景の疑似要素 */
-.tooltip_background::after{
-  content: "";
-  position: absolute;
-  z-index: 2;
-  width: 100%;
-  height:100%;
-  top: 0;
-  left: 0;
-  background-color: white; /* 背景の色 */
-  opacity: 0.7; /* 0.0 ~ 1.0　小さくなるほど透明になる */
-}
-
-.hover:hover {
-    outline: 1px solid #000;
-    outline-offset: -1px;
-}
