--- TrainingList/src/main/resources/templates/teacher/teacherinfo.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-
-<head th:replace="common/head :: head_fragment(title = '講師割当アプリケーション [講師情報]', scripts = ~{::script}, links = ~{::link})"></head>
-<body>
-<div class="container">
-  <br>
-  <h2>講師情報</h2>
-  <br>
-  <br>
-  <form th:action="@{/teacher/create}" th:object="${teacherRequest}" th:method="post">
-    <div class="input-group mb-3" hidden>
-      <span class="input-group-text">ID</span>
-      <input type="text" class="form-control" th:field="*{id}" />
-    </div>
-    <div class="text-danger" th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></div>
-    <div class="input-group mb-3">
-      <span class="input-group-text">講師名</span>
-      <input type="text" class="form-control" th:field="*{name}" />
-    </div>
-
-    <br>
-    <div class="text-center">
-      <input type="submit" value="　登録　" class="btn btn-primary btn-block">
-      <a href="/teacher/list" class="btn btn-outline-primary btn-block">キャンセル</a>
-    </div>
-  </form>
-</div>
-</body>
-</html>
