--- TrainingList/src/main/resources/templates/training/editschedule.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-
-<head th:replace="common/head :: head_fragment(title = '講師割当アプリケーション [スケジュール登録]', scripts = ~{::script}, links = ~{::link})"></head>
-<body>
-<div class="container">
-<br>
-<br>
-<br>
-<form th:action="@{/schedule/edit}" th:object="${schedulelist}" th:method="post">
-    <nav class="navbar fixed-top bg-body-tertiary">
-        <div class="container-fluid">
-            <a class="navbar-brand">
-                <h5><span th:text="${trainingname}"></span></h5>
-            </a>
-            <div class="text-end">
-                <input type="submit" value="　更新　" class="btn btn-primary btn-block" />
-                <input type="hidden" th:field="*{training_id}" />
-                <!--/* <a th:href="@{/training/timetable/{id}(id=${training_id})}" class="btn btn-outline-primary btn-block">キャンセル</a> */-->
-            </div>
-        </div>
-    </nav>
-
-    <table class="table table-striped table-hover">
-        <thead class="thead-dark">
-        <tr>
-            <th>訓練日</th>
-            <th>訓練時間</th>
-            <th>講師</th>
-            <th>備考</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="row, st : *{scheduleDataList}">
-            <td hidden>
-                <input type="hidden" th:field="*{scheduleDataList[__${st.index}__].id}">
-                <input type="hidden" th:field="*{scheduleDataList[__${st.index}__].training_id}">
-            </td>
-            <td>
-                <label th:text="${#temporals.format(row.training_date, 'yyyy/MM/dd(eee)')}"></label>
-                <input type="hidden" th:field="*{scheduleDataList[__${st.index}__].training_date}">
-            </td>
-            <td>
-                <input type="number" class="form-control" min="0" max="8"  th:field="*{scheduleDataList[__${st.index}__].training_hours}" />
-            </td>
-            <td>
-                <select class="form-select" th:field="*{scheduleDataList[__${st.index}__].teacher_id}">
-                    <option value="0"></option>
-                    <option
-                            th:each="teacher : ${teacherlist}"
-                            th:value="${teacher.id}"
-                            th:text="${teacher.name}">
-                    </option>
-                    <!--/* th:selected="${teacher.id} == *{scheduleDataList[__${st.index}__].teacher_id}" */-->
-                </select>
-            </td>
-            <td>
-                <input type="text" class="form-control" th:field="*{scheduleDataList[__${st.index}__].memo}">
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</form>
-</div>
-<br>
-<br>
-<!-- Navbar -->
-<nav class="navbar fixed-bottom navbar-expand-sm navbar-dark bg-dark">
-    <a th:href="@{/training/list}" class="btn btn-sm btn-outline-light btn-sm">訓練一覧へ</a>
-    <a th:href="@{/training/timetable/{id}(id=${schedulelist.training_id})}" class="btn btn-sm btn-outline-light">訓練時間表へ</a>
-</nav>
-</body>
-</html>
