--- TrainingList/src/main/resources/templates/training/traininginfo.html
@@ -1,87 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-
-<head th:replace="common/head :: head_fragment(title = '講師割当アプリケーション [訓練情報]', scripts = ~{::script}, links = ~{::link})"></head>
-
-<body>
-	<div class="container">
-		<br>
-		<h2>訓練情報</h2>
-		<br>
-		<!--/* 新規登録の場合「訓練スケジュール登録」ボタン非表示 */-->
-		<th:block th:if="${trainingRequest.id != 0}">
-			<div class="text-end">
-				<a th:href="@{/training/editschedule/{id}(id=${trainingRequest.id})}" class="btn btn-success">訓練スケジュール登録</a>
-			</div>
-		</th:block>
-		<br>
-		<form th:action="@{/training/create}" th:object="${trainingRequest}" th:method="post">
-			<div class="input-group mb-3" hidden>
-				<span class="input-group-text">ID</span>
-				<input type="text" class="form-control" th:field="*{id}" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('trainingname')}" th:errors="*{trainingname}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">訓練名</span>
-				<input type="text" class="form-control" th:field="*{trainingname}" th:errorclass="is-invalid" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('dateValidRecruit')}" th:errors="*{dateValidRecruit}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('recruit_start_date')}" th:errors="*{recruit_start_date}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('recruit_end_date')}" th:errors="*{recruit_end_date}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">募集期間</span>
-				<input type="date" class="form-control" name="recruit_start_date" th:value="*{recruit_start_date}" th:errorclass="is-invalid" />　～　
-				<input type="date" class="form-control" name="recruit_end_date" th:value="*{recruit_end_date}" th:errorclass="is-invalid" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('selection_date')}" th:errors="*{selection_date}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">選考日</span>
-				<input type="date" class="form-control" name="selection_date" th:value="*{selection_date}" th:errorclass="is-invalid" />
-			</div>
-<!--			<div class="text-danger" th:if="${#fields.hasErrors('acceptance_date')}" th:errors="*{acceptance_date}"></div>-->
-			<div class="input-group mb-3">
-				<span class="input-group-text">合格発表</span>
-				<input type="date" class="form-control" name="acceptance_date" th:value="*{acceptance_date}" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('dateValidTraining')}" th:errors="*{dateValidTraining}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('start_date')}" th:errors="*{start_date}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('end_date')}" th:errors="*{end_date}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">訓練期間</span>
-				<input type="date" class="form-control" name="start_date" th:value="*{start_date}" th:errorclass="is-invalid" />　～　
-				<input type="date" class="form-control" name="end_date" th:value="*{end_date}" th:errorclass="is-invalid" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('timeValidTraining')}" th:errors="*{timeValidTraining}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('start_time')}" th:errors="*{start_time}"></div>
-			<div class="text-danger" th:if="${#fields.hasErrors('end_time')}" th:errors="*{end_time}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">授業時間</span>
-				<input type="time" list="data-list" min="09:00" max="13:00" class="form-control" name="start_time" th:value="*{start_time}" th:errorclass="is-invalid" />　～　
-				<input type="time" list="data-list" min="11:00" max="18:00" class="form-control" name="end_time" th:value="*{end_time}" th:errorclass="is-invalid" />
-			</div>
-			<div class="text-danger" th:if="${#fields.hasErrors('training_hours')}" th:errors="*{training_hours}"></div>
-			<div class="input-group mb-3">
-				<span class="input-group-text">授業時間数</span>
-				<input type="number" class="form-control" min="1" max="8" name="training_hours" th:value="*{training_hours}" th:errorclass="is-invalid" />
-			</div>
-			<br>
-			<div class="text-center">
-				<input type="submit" value="　登録　" class="btn btn-primary btn-block">
-				<a href="/training/list" class="btn btn-outline-primary btn-block">キャンセル</a>
-			</div>
-		</form>
-    </div>
-	<datalist id="data-list">
-		<option value="09:00"></option><option value="09:10"></option><option value="09:20"></option><option value="09:30"></option><option value="09:40"></option><option value="09:50"></option>
-		<option value="10:00"></option><option value="10:10"></option><option value="10:20"></option><option value="10:30"></option><option value="10:40"></option><option value="10:50"></option>
-		<option value="11:00"></option><option value="11:10"></option><option value="11:20"></option><option value="11:30"></option><option value="11:40"></option><option value="11:50"></option>
-		<option value="12:00"></option><option value="12:10"></option><option value="12:20"></option><option value="12:30"></option><option value="12:40"></option><option value="12:50"></option>
-		<option value="13:00"></option><option value="13:10"></option><option value="13:20"></option><option value="13:30"></option><option value="13:40"></option><option value="13:50"></option>
-		<option value="14:00"></option><option value="14:10"></option><option value="14:20"></option><option value="14:30"></option><option value="14:40"></option><option value="14:50"></option>
-		<option value="15:00"></option><option value="15:10"></option><option value="15:20"></option><option value="15:30"></option><option value="15:40"></option><option value="15:50"></option>
-		<option value="16:00"></option><option value="16:10"></option><option value="16:20"></option><option value="16:30"></option><option value="16:40"></option><option value="16:50"></option>
-		<option value="17:00"></option><option value="17:10"></option><option value="17:20"></option><option value="17:30"></option><option value="17:40"></option><option value="17:50"></option>
-		<option value="18:00"></option>
-	</datalist>
-</body>
-</html>
