--- TrainingList/src/main/resources/templates/training/traininglist.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="ja">
-
-<head th:replace="common/head :: head_fragment(title = '講師割当アプリケーション [訓練一覧]', scripts = ~{::script}, links = ~{::link})"></head>
-
-<body>
-<div class="container">
-    <br>
-    <h2>訓練一覧</h2>
-    <div class="float-end">
-        <a th:href="@{/training/add}" class="btn btn-primary">新規登録</a>
-    </div>
-    <table class="table table-striped table-hover">
-        <thead class="thead-dark">
-        <tr>
-            <th>訓練名</th>
-            <th>訓練期間</th>
-            <th></th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="training : ${traininglist}"  th:object="${training}">
-            <td th:text="*{id}" hidden></td>
-            <!--/* <td th:text="*{trainingname}"></td> */-->
-            <td><a th:href="@{/training/timetable/{id}(id=*{id})}" th:text="*{trainingname}"></a></td>
-            <td th:text="*{#temporals.format(start_date, 'yyyy/MM/dd(eee)')} + ' ～ ' + *{#temporals.format(end_date, 'yyyy/MM/dd(eee)')}"></td>
-            <td>
-                <!--/* <a th:href="@{/training/timetable/{id}(id=*{id})}" class="btn btn-outline-secondary btn-sm">スケジュール</a> */-->
-                <a th:href="@{/training/update/{id}(id=*{id})}" class="btn btn-outline-secondary btn-sm">編集</a>
-                <a th:href="@{/training/delete/{id}(id=*{id})}" class="btn btn-outline-secondary btn-sm" onclick="return confirm_delete(this.parentElement.parentElement.cells[1].firstChild.data)">削除</a>
-            </td>
-        </tbody>
-    </table>
-</div>
-<br>
-<br>
-<!-- Navbar -->
-<nav class="navbar fixed-bottom navbar-expand-sm navbar-dark bg-dark">
-    <a th:href="@{/teacher/list}" class="btn btn-sm btn btn-outline-light">講師一覧へ</a>
-</nav>
-<script>
-    <!--/* 削除確認メッセージ */-->
-	function confirm_delete(data) {
-		return window.confirm(data + "を削除しますか？")
-    };
-</script>
-</body>
-</html>
