--- TrainingList/src/test/java/jp/crossabilitys/work/TrainingList/TrainingListApplicationTests.java
@@ -1,13 +1,0 @@
-package jp.crossabilitys.work.TrainingList;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TrainingListApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
