+++ src/main/resources/static/timetable.js
